% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancerdt2_1}
\alias{cancerdt2_1}
\title{Example omics data for this package}
\format{
An object of class \code{data.frame} with 194 rows and 16051 columns.
}
\source{
curatedOvarianData.
}
\usage{
data(exampledt, package = 'SurvBenchmark')
}
\description{
Example omics data for this package
}
\keyword{datasets}
