% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_cox_boost_fun.R
\name{ga_cox_boost_fun}
\alias{ga_cox_boost_fun}
\title{CoxBoost model using GA as feature selection}
\usage{
ga_cox_boost_fun(
  r,
  data,
  cvK,
  numm,
  topnumm,
  generation_num,
  time1,
  timess,
  stepnumber,
  penaltynumber
)
}
\arguments{
\item{r}{a numeric value, a seed to run this method}

\item{data}{a dataframe, the data used to performance this survival model}

\item{cvK}{a numeric value, cross-validation fold}

\item{numm}{a numeric value, the number of variables,i.e.for example, number of proteins in the data}

\item{topnumm}{a numeric value, the number of variables selected to be passed into the model, for example, the number of DE genes}

\item{generation_num}{a numeric value, the generation number used in the GA algorithm, details see package "GenAlgo"}

\item{time1}{a numeric value, the time point to calculate the risk, see package "CoxBoost"}

\item{timess}{a numeric vector of length 15, contains time points to get the time-dependent AUC values}

\item{stepnumber}{a numeric value, the number of stpes performed in the model, see package "CoxBoost"}

\item{penaltynumber}{a numeric value, the penalty number used in the model, see package "CoxBoost"}
}
\value{
a data.frame with allevaluation measurements in all columns and rows are each fold results from cross-validation
}
\description{
CoxBoost model using GA as feature selection
}
\examples{
data("exampledt", package = "SurvBenchmark")
fitform_ogl=survival::Surv(time,status)~.
formula1=fitform_ogl
formula2=fitform_ogl
formula3=survival::Surv(time,status)~1
formula4=survival::Surv(time,status)~1
form1=as.formula(~.)
timess=seq(as.numeric(summary(cancerdt2_1$time)[2]),as.numeric(summary(cancerdt2_1$time)[5]),(as.numeric(summary(cancerdt2_1$time)[5])-as.numeric(summary(cancerdt2_1$time)[2]))/14)
time1=timess[3]
stepnumber=10
penaltynumber=100
want=ga_cox_boost_fun(1,cancerdt2_1,5, 16047,5,20,time1,timess,stepnumber,penaltynumber);
}
