% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_mtlr_fun.R
\name{ga_mtlr_fun}
\alias{ga_mtlr_fun}
\title{mtlr model using GA as feature selection}
\usage{
ga_mtlr_fun(r, data, cvK, numm, topnumm, generation_num, timess)
}
\arguments{
\item{r}{a numeric value, a seed to run this method}

\item{data}{a dataframe, the data used to performance this survival model}

\item{cvK}{a numeric value, cross-validation fold}

\item{numm}{a numeric value, the number of variables,i.e.for example, number of proteins in the data}

\item{topnumm}{a numeric value, the number of variables selected to be passed into the model, for example, the number of DE genes}

\item{generation_num}{a numeric value, the generation number used in the GA algorithm, details see package "GenAlgo"}

\item{timess}{a numeric vector, contains time points to get the time-dependent AUC values}
}
\value{
a data.frame with allevaluation measurements in all columns and rows are each fold results from cross-validation
}
\description{
mtlr model using GA as feature selection
}
\examples{
data("exampledt", package = "SurvBenchmark")
fitform_ogl=survival::Surv(time,status)~.
formula1=fitform_ogl
formula2=fitform_ogl
formula3=survival::Surv(time,status)~1
formula4=survival::Surv(time,status)~1
form1=as.formula(~.)
timess=seq(as.numeric(summary(cancerdt2_1$time)[2]),as.numeric(summary(cancerdt2_1$time)[5]),(as.numeric(summary(cancerdt2_1$time)[5])-as.numeric(summary(cancerdt2_1$time)[2]))/14)
want=ga_mtlr_fun(1,cancerdt2_1,5, 16047,5,20,timess);
}
