% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnsurv_fun.R
\name{getPseudoConditional}
\alias{getPseudoConditional}
\title{dnnsurv model}
\usage{
getPseudoConditional(t, d, qt)
}
\arguments{
\item{r}{a numeric value, a seed to run this method}

\item{data}{a dataframe, the data used to performance this survival model}

\item{cvK}{a numeric value, cross-validation fold}

\item{fitform_ogl}{a Surv object from package survival, the survival function}

\item{formula1}{a Surv object from package survival, to calculate a version of the brier score, details please check package pec}

\item{formula2}{a Surv object from package survival, to calculate a version of the brier score, details please check package pec}

\item{formula3}{a Surv object from package survival, to calculate a version of the brier score, details please check package pec}

\item{formula4}{a Surv object from package survival, to calculate a version of the brier score, details please check package pec}

\item{timess}{a numeric vector, contains time points to get the time-dependent AUC values}

\item{pickedtime}{a numeric vector, the picked pesudo time points, see package "DNNSurv" for details}
}
\value{
a data.frame with all evaluation measurements in all columns and rows are each fold results from cross-validation
}
\description{
dnnsurv model
}
\examples{
data("exampledt2", package = "SurvBenchmark")
xnam <- paste(colnames(veteran)[c(1,2,5,6,7,8)], sep="")
form=as.formula(paste("survival::Surv(time, status)~ ", paste(xnam, collapse= "+")))
fitform_ogl=form
formula1=fitform_ogl
formula2=fitform_ogl
formula3=survival::Surv(time,status)~1
formula4=survival::Surv(time,status)~1
pickedtime=c(2,3,5,7,9)
timess=seq(as.numeric(summary(veteran$time)[2]),as.numeric(summary(veteran$time)[5]),(as.numeric(summary(veteran$time)[5])-as.numeric(summary(veteran$time)[2]))/14)
want=dnnsurv_fun(1,veteran,5,fitform_ogl,formula1,formula2,formula3,formula4,timess,pickedtime);
}
