% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_ridge_fun.R
\name{glmnet_ridge_fun}
\alias{glmnet_ridge_fun}
\title{ridge cox model using glmnet}
\usage{
glmnet_ridge_fun(r, data, cvK, formula1, formula2, formula3, formula4, timess)
}
\arguments{
\item{r}{a numeric value, a seed to run this method}

\item{data}{a dataframe, the data used to performance this survival model}

\item{cvK}{a numeric value, cross-validation fold}

\item{formula1}{a Surv object from package survival, to calculate a version of the brier score, details please check package pec}

\item{formula2}{a Surv object from package survival, to calculate a version of the brier score, details please check package pec}

\item{formula3}{a Surv object from package survival, to calculate a version of the brier score, details please check package pec}

\item{formula4}{a Surv object from package survival, to calculate a version of the brier score, details please check package pec}

\item{timess}{a numeric vector of length 15, contains time points to get the time-dependent AUC values}

\item{numm}{a numeric value, the number of variables,i.e.for example, number of proteins in the data}

\item{topnumm}{a numeric value, the number of variables selected to be passed into the model, for example, the number of DE genes}

\item{fitform_ogl}{a Surv object from package survival, the survival function}
}
\value{
a data.frame with allevaluation measurements in all columns and rows are each fold results from cross-validation
}
\description{
ridge cox model using glmnet
}
\examples{
data("exampledt", package = "SurvBenchmark")
fitform_ogl=survival::Surv(time,status)~.
formula1=fitform_ogl
formula2=fitform_ogl
formula3=survival::Surv(time,status)~1
formula4=survival::Surv(time,status)~1
form1=as.formula(~.)
timess=seq(as.numeric(summary(cancerdt2_1$time)[2]),as.numeric(summary(cancerdt2_1$time)[5]),(as.numeric(summary(cancerdt2_1$time)[5])-as.numeric(summary(cancerdt2_1$time)[2]))/14)
want=glmnet_ridge_fun(1,cancerdt2_1[,-dim(cancerdt2_1)[2]],5,formula1,formula2,formula3,formula4,timess);
}
