% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{veteran}
\alias{veteran}
\title{Example clinical data for this package}
\format{
An object of class \code{data.frame} with 137 rows and 8 columns.
}
\source{
veteran data.
}
\usage{
data(exampledt2, package = 'SurvBenchmark')
}
\description{
Example clinical data for this package
}
\keyword{datasets}
